requirements.txt
aiogram~=2.16
python-binance
finplot